# Conflicts:
#	code/__HELPERS/logging.dm
#	maps/CEVEris/_CEV_Eris.dmm

Brings branch back up to date, expect a lot of files to be changed

+++ code/game/objects/items/weapons/melee/energy.dm
@@ -1,8 +1,4 @@
 /obj/item/weapon/melee/energy
-	var/active = 0
-	var/active_force
-	var/active_throwforce
-	var/active_w_class
 	sharp = FALSE
 	edge = FALSE
 	armor_penetration = ARMOR_PEN_HALF
@@ -10,6 +6,11 @@
 	structure_damage_factor = STRUCTURE_DAMAGE_BREACHING
 	heat = 3800
 	embed_mult = 0 //No physical matter to catch onto things
+	bad_type = /obj/item/weapon/melee/energy
+	var/active = 0
+	var/active_force
+	var/active_throwforce
+	var/active_w_class
 
 /obj/item/weapon/melee/energy/is_hot()
 	if (active)
@@ -29,7 +30,7 @@
 	update_wear_icon()
 
 /obj/item/weapon/melee/energy/proc/deactivate(mob/living/user)
-	anchored = 0
+	anchored = FALSE
 	if(!active)
 		return
 	playsound(user, 'sound/weapons/saberoff.ogg', 50, 1)
@@ -109,25 +110,27 @@
 	..()
 	deactivate(user)
 
-/obj/item/weapon/melee/energy/sword/New()
-	blade_color = pick("red","blue","green","purple")
+/obj/item/weapon/melee/energy/sword/Initialize(mapload)
+	. = ..()
+	if(!blade_color)
+		blade_color = pick("red","blue","green","purple")
 
-/obj/item/weapon/melee/energy/sword/green/New()
+/obj/item/weapon/melee/energy/sword/green
 	blade_color = "green"
 
-/obj/item/weapon/melee/energy/sword/red/New()
+/obj/item/weapon/melee/energy/sword/red
 	blade_color = "red"
 
-/obj/item/weapon/melee/energy/sword/blue/New()
+/obj/item/weapon/melee/energy/sword/blue
 	blade_color = "blue"
 
-/obj/item/weapon/melee/energy/sword/purple/New()
+/obj/item/weapon/melee/energy/sword/purple
 	blade_color = "purple"
 
-/obj/item/weapon/melee/energy/sword/pirate/New()
+/obj/item/weapon/melee/energy/sword/pirate
 	blade_color = "cutlass"
-	
-/obj/item/weapon/melee/energy/sword/sabre/New()
+
+/obj/item/weapon/melee/energy/sword/sabre
 	blade_color = "green"
 
 /obj/item/weapon/melee/energy/sword/activate(mob/living/user)
@@ -192,14 +195,16 @@
 	attack_verb = list("attacked", "slashed", "stabbed", "sliced", "torn", "ripped", "diced", "cut")
 	var/mob/living/creator
 	var/datum/effect/effect/system/spark_spread/spark_system
+	var/cleanup = TRUE	// Should the blade despawn moments after being discarded by the summoner?
+	var/init_procees = TRUE
 
-/obj/item/weapon/melee/energy/blade/New()
-
-	spark_system = new /datum/effect/effect/system/spark_spread()
-	spark_system.set_up(5, 0, src)
-	spark_system.attach(src)
-
-	START_PROCESSING(SSobj, src)
+/obj/item/weapon/melee/energy/blade/Initialize(mapload)
+	. = ..()
+	if(init_procees)
+		spark_system = new /datum/effect/effect/system/spark_spread()
+		spark_system.set_up(5, 0, src)
+		spark_system.attach(src)
+		START_PROCESSING(SSobj, src)
 
 /obj/item/weapon/melee/energy/blade/Destroy()
 	STOP_PROCESSING(SSobj, src)
@@ -207,10 +212,12 @@
 
 /obj/item/weapon/melee/energy/blade/attack_self(mob/user as mob)
 	user.drop_from_inventory(src)
-	spawn(1) if(src) qdel(src)
+	if(cleanup)
+		spawn(1) if(src) qdel(src)
 
 /obj/item/weapon/melee/energy/blade/dropped()
-	spawn(1) if(src) qdel(src)
+	if(cleanup)
+		spawn(1) if(src) qdel(src)
 
 /obj/item/weapon/melee/energy/blade/Process()
 	if(!creator || loc != creator || (creator.l_hand != src && creator.r_hand != src))
@@ -225,4 +232,11 @@
 			host.pinned -= src
 			host.embedded -= src
 			host.drop_from_inventory(src)
-		spawn(1) if(src) qdel(src)
+		if(cleanup)
+			spawn(1) if(src) qdel(src)
+
+/obj/item/weapon/melee/energy/blade/organ_module //just to make sure that blade doesnt delet itself
+	cleanup = FALSE
+	init_procees = FALSE
+
+/obj/item/weapon/melee/energy/blade/organ_module/attack_self(mob/user)
