+++ code/modules/projectiles/guns/projectile/shotgun/doublebarrel.dm
@@ -11,12 +11,13 @@
 	max_shells = 2
 	w_class = ITEM_SIZE_HUGE
 	force = WEAPON_FORCE_PAINFUL
-	flags =  CONDUCT
+	flags = CONDUCT
 	slot_flags = SLOT_BACK
 	caliber = CAL_SHOTGUN
+	recoil_buildup = 7
 	origin_tech = list(TECH_COMBAT = 3, TECH_MATERIAL = 1)
 	ammo_type = /obj/item/ammo_casing/shotgun/beanbag
-	bulletinsert_sound 	= 'sound/weapons/guns/interact/shotgun_insert.ogg'
+	bulletinsert_sound = 'sound/weapons/guns/interact/shotgun_insert.ogg'
 	fire_sound = 'sound/weapons/guns/fire/shotgunp_fire.ogg'
 	matter = list(MATERIAL_PLASTEEL = 20, MATERIAL_WOOD = 10)
 	price_tag = 1200
@@ -24,8 +25,8 @@
 	var/bolt_open = 0
 	burst_delay = 0
 	init_firemodes = list(
-		list(mode_name="fire one barrel at a time", burst=1, icon="semi"),
-		list(mode_name="fire both barrels at once", burst=2, icon="burst"),
+		list(mode_name="Single-fire", mode_desc="Send Vagabonds flying back several paces", burst=1, icon="semi"),
+		list(mode_name="Both Barrels", mode_desc="Give them the side-by-side", burst=2, icon="burst"),
 		)
 
 /obj/item/weapon/gun/projectile/shotgun/doublebarrel/pellet
@@ -84,16 +85,22 @@
 
 /obj/item/weapon/gun/projectile/shotgun/doublebarrel/attackby(var/obj/item/A as obj, mob/user as mob)
 	if(QUALITY_SAWING in A.tool_qualities)
-		to_chat(user, SPAN_NOTICE("You begin to shorten the barrel of \the [src]."))
-		if(loaded.len)
-			for(var/i in 1 to max_shells)
-				afterattack(user, user)	//will this work? //it will. we call it twice, for twice the FUN
-				playsound(user, fire_sound, 50, 1)
-			user.visible_message(SPAN_DANGER("The shotgun goes off!"), SPAN_DANGER("The shotgun goes off in your face!"))
-			return
-		if(A.use_tool(user, src, WORKTIME_FAST, QUALITY_SAWING, FAILCHANCE_NORMAL, required_stat = STAT_COG))
-			qdel(src)
-			new /obj/item/weapon/gun/projectile/shotgun/doublebarrel/sawn(usr.loc)
-			to_chat(user, SPAN_WARNING("You shorten the barrel of \the [src]!"))
+		if (!istype(src, /obj/item/weapon/gun/projectile/shotgun/doublebarrel/sawn))
+			if (src.item_upgrades.len)
+				if("No" == input(user, "There are attachments present. Would you like to destroy them?") in list("Yes", "No"))
+					return
+			to_chat(user, SPAN_NOTICE("You begin to shorten the barrel of \the [src]."))
+			if(loaded.len)
+				for(var/i in 1 to max_shells)
+					afterattack(user, user)	//will this work? //it will. we call it twice, for twice the FUN
+					playsound(user, fire_sound, 50, 1)
+				user.visible_message(SPAN_DANGER("The shotgun goes off!"), SPAN_DANGER("The shotgun goes off in your face!"))
+				return
+			if(A.use_tool(user, src, WORKTIME_FAST, QUALITY_SAWING, FAILCHANCE_NORMAL, required_stat = STAT_COG))
+				qdel(src)
+				new /obj/item/weapon/gun/projectile/shotgun/doublebarrel/sawn(usr.loc)
+				to_chat(user, SPAN_WARNING("You shorten the barrel of \the [src]!"))
+		else
+			to_chat(user, SPAN_WARNING("You cannot shorten \the [src] any further!"))
 	else
 		..()
